feat: proteger rutas y endpoints admin

+++ backend/app/routes/category_routes.py
@@ -1,7 +1,9 @@
 from flask import Blueprint, jsonify, request
+from flask_jwt_extended import get_jwt, jwt_required
 
 from app.models.exceptions import CampoIncorrectoException, CategoriaConProductosException, CategoriaYaExistenteException, CategoriaNoEncontradaException
 from app.services.category_services import obtener_todas_categorias, insertar_categoria_base, eliminar_categoria_base, actualizar_categoria
+from app.services.auth_services import comprobar_usuario_es_admin
 
 from app.models.categoria import Categoria
 
@@ -18,7 +20,13 @@ def get_categorias():
 
 # POST INSERTAR CATEGORÍA
 @categoria_bp.route("/", methods=["POST"])
+@jwt_required()
 def post_categoria():
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     data = request.get_json()
 
     if not data or not data.get("nombre") or not data.get("descripcion"):
@@ -39,7 +47,13 @@ def post_categoria():
 
 # DELETE CATEGORÍA
 @categoria_bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def del_categoria(id):
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     try:
         eliminar_categoria_base(id)
 
@@ -52,7 +66,13 @@ def del_categoria(id):
 
 # EDITAR CATEGORÍA
 @categoria_bp.route("/<int:id>", methods=["PUT"])
+@jwt_required()
 def put_categoria(id):
+    data = get_jwt()
+
+    if comprobar_usuario_es_admin(data) is False:
+        return jsonify({"error": "Acción cancelada. El usuario no es admin"}), 403
+
     data = request.get_json()
 
     if not data or not data.get("nombre") or not data.get("descripcion"):
